BIOS: Properly read sector count for big SATA disks.
svn:sync-xref-src-repo-rev: r80099

+++ src/VBox/Devices/PC/BIOS/ahci.c
@@ -1,4 +1,4 @@
-/* $Id: ahci.c 42811 2012-08-14 16:00:48Z michal.necasek@oracle.com $ */
+/* $Id: ahci.c 42842 2012-08-16 11:08:49Z michal.necasek@oracle.com $ */
 /** @file
  * AHCI host adapter driver to boot from SATA disks.
  */
@@ -668,7 +668,7 @@ static void ahci_port_detect_device(ahci_t __far *ahci, uint8_t u8Port)
 
                 /** @todo update sectors to be a 64 bit number (also lba...). */
                 if (cSectors == 268435455)
-                    cSectors = *(uint16_t *)(abBuffer+(100*2)); // words 100 to 103 (someday)
+                    cSectors = *(uint32_t *)(abBuffer+(100*2)); // words 100 to 103 (someday)
 
                 DBG_AHCI("AHCI: %ld sectors\n", cSectors);
